# feat: improve split performance