reduce construction of regexp and replacer.

```
name        old time/op  new time/op  delta
ToGoName-4  1.99ms ± 3%  0.44ms ± 2%  -77.98%  (p=0.000 n=10+10)
```