@casualjim 2018-11-16
We have to change the golang checks for this because I think the use case is valid